pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quickpush"
version = "0.1.0"
description = "A helper tool for faster PR creation"
requires-python = ">=3.8"
dependencies = [
    "inquirer==3.1.3",
    "PyGithub==2.1.1",
]

[project.scripts]
quickpush = "quickpush:main" 